backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
pymysql==1.1.0
psycopg2-binary==2.9.9
oracledb==2.0.0
pymssql==2.2.11
pycryptodome==3.20.0
python-dotenv==1.0.0
